Comments added in migrations

+++ database/migrations/2021_07_07_085546_create_exams_table.php
@@ -18,7 +18,7 @@ class CreateExamsTable extends Migration
             $table->foreignId("ssd_id")->nullable()->constrained()->cascadeOnDelete();
             $table->string("code")->unique()->nullable(); //unused right now, but left for future applications
             $table->string("name");
-            $table->integer("cfu");
+            $table->unsignedTinyInteger("cfu");
             $table->timestamps();
         });
     }
